+++ run.py
@@ -1,0 +1,6 @@
+def main():
+    pass
+
+
+if '__main__' in __name__:
+    main()
