+++ tests/test_json_file.py
@@ -1,10 +1,9 @@
 import unittest
 from unittest import TestCase
 import os
-import jsonschema
 import json
 
-from pisa_to_json import main
+from pisa_utils import analyse
 
 test_data_path= os.path.join('tests', 'data')
 
@@ -18,7 +17,7 @@ class TestJsonFile(TestCase):
 
         pisa_config_file=os.path.join(test_data_path,'pisa.cfg')
 
-        ap= main.AnalysePisa(pdbid_id="6nxr", assembly_id="1", pisa_config = pisa_config_file,output_dir=test_data_path,force=True,result_json_file="output.json",input_dir=test_data_path)
+        ap= analyse.AnalysePisa(pdbid_id="6nxr", assembly_id="1", pisa_config = pisa_config_file, output_dir=test_data_path, force=True, result_json_file="output.json", input_dir=test_data_path)
         
         ap.run_process()
