+++ tests/test_xml_files.py
@@ -3,7 +3,7 @@ from unittest import TestCase
 import os
 import xmltodict
 
-from pisa_to_json import main
+from pisa_utils import analyse
 test_data_path= os.path.join('tests', 'data')
 
 class TestXmlFiles(TestCase):
@@ -25,7 +25,7 @@ class TestXmlFiles(TestCase):
 
         pisa_config_file=os.path.join(test_data_path,'pisa.cfg')
 
-        ap= main.AnalysePisa(pdbid_id="6nxr", assembly_id="1", pisa_config = pisa_config_file, output_dir=test_data_path,force=True,result_json_file="output.json",input_dir=test_data_path)
+        ap= analyse.AnalysePisa(pdbid_id="6nxr", assembly_id="1", pisa_config = pisa_config_file, output_dir=test_data_path, force=True, result_json_file="output.json", input_dir=test_data_path)
         
         ap.run_process()
